# core-module/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Национальный сервис мониторинга цен на продовольственные товары(OpenAPI 3.0)
  description: |-
    Национальный сервис мониторинга цен на продовольственные товары
    
    Полезные ссылки:
    - [Репозиторий](https://github.com/AnnaDesert/Learning)
  version: 1.0.11
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /shop:
    get:
      tags:
        - shop
      summary: Получить все торговые точки
      description: Получить все торговые точки
      operationId: allShop
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShopDTO'
    post:
      tags:
        - shop
      summary: Добваить торговую точку
      description: Добваить торговую точку
      operationId: addShop
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopCreationDTO'
        required: true
      responses:
        '201':
          description: Created
  /shop/{id}:
    get:
      tags:
        - shop
      summary: Получить торговую точку по ID
      description: Получить торговую точку по ID
      operationId: getShopById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ShopDTO'
    delete:
      tags:
        - shop
      summary: Удалить торговую точку по ID
      description: Удалить торговую точку по ID
      operationId: removeShopById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
    patch:
      tags:
        - shop
      summary: Изменить торговую точку
      description: Изменить торговую точку
      operationId: updateShop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopCreationDTO'
        required: true
      responses:
        '200':
          description: OK
  /shop/name/{name}:
    get:
      tags:
        - shop
      summary: Получить торговую точку по названию
      description: Получить торговую точку по названию
      operationId: getShopByName
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShopDTO'
  /shop/address/{address}:
    get:
      tags:
        - shop
      summary: Получить торговую точку по адресу
      description: Получить торговую точку по адресу
      operationId: getShopByAddress
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShopDTO'
  /product:
    get:
      tags:
        - product
      summary: Получить все товары
      description: Получить все товары
      operationId: allProducts
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDTO'
    post:
      tags:
        - product
      summary: Добваить продукт
      description: Добваить новый продукт
      operationId: addProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreationDTO'
        required: true
      responses:
        '201':
          description: Created
  /product/{id}:
    get:
      tags:
        - product
      summary: Получить товар по ID
      description: Получить товар по ID
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDTO'
    delete:
      tags:
        - product
      summary: Удалить товар по ID
      description: Удалить товар по ID
      operationId: removeProductById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
    patch:
      tags:
        - product
      summary: Изменить продукт
      description: Изменить продукт
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreationDTO'
        required: true
      responses:
        '200':
          description: OK
  /product/name/{name}:
    get:
      tags:
        - product
      summary: Получить товар по названию
      description: Получить товар по названию
      operationId: getProductByName
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDTO'
  /product/category/{category}:
    get:
      tags:
        - product
      summary: Получить товар по категории
      description: Получить товар по категории
      operationId: getProductByCategory
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDTO'
  /product/{fileName}:
    post:
      tags:
        - product
      summary: Добавить товары из csv файла
      description: Добавить товары из csv файла
      operationId: addProductFromCSV
      parameters:
        - name: fileName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
...
components:
  schemas:
...
    ShopCreationDTO:
      required:
        - address
        - idAdmin
        - name
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        idAdmin:
          type: integer
          format: int64
...
    ProductCreationDTO:
      required:
        - name
      type: object
      properties:
        name:
          type: string
...
    ShopDTO:
      required:
        - address
        - id
        - idAdmin
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        address:
          type: string
        idAdmin:
          type: integer
          format: int64
    ProductDTO:
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...