# user-module/target/generated-sources/swagger/main/java/org/senla/swagger/api/ProductApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.55).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package org.senla.swagger.api;

import org.senla.swagger.model.ProductCreationDTO;
import org.senla.swagger.model.ProductDTO;
import com.fasterxml.jackson.databind.ObjectMapper;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.servlet.http.HttpServletRequest;
import javax.validation.Valid;
import javax.validation.constraints.*;
import java.io.IOException;
import java.util.List;
import java.util.Map;
import java.util.Optional;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2024-06-24T09:46:33.073075700+03:00[Europe/Moscow]")
@Validated
public interface ProductApi {

    Logger log = LoggerFactory.getLogger(ProductApi.class);

    Optional<ObjectMapper> getObjectMapper();

    Optional<HttpServletRequest> getRequest();

    

    @Operation(summary = "Добваить продукт", description = "Добваить новый продукт", tags={ "product" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "201", description = "Created") })
    @RequestMapping(value = "/product",
        consumes = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<Void> addProduct(@Parameter(in = ParameterIn.DEFAULT, description = "", required=true, schema=@Schema()) @Valid @RequestBody ProductCreationDTO body
);


    @Operation(summary = "Добавить товары из csv файла", description = "Добавить товары из csv файла", tags={ "product" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "OK") })
    @RequestMapping(value = "/product/{fileName}",
        method = RequestMethod.POST)
    ResponseEntity<Void> addProductFromCSV(@Parameter(in = ParameterIn.PATH, description = "", required=true, schema=@Schema()) @PathVariable("fileName") String fileName
);


    @Operation(summary = "Получить все товары", description = "Получить все товары", tags={ "product" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "OK", content = @Content(mediaType = "*/*", array = @ArraySchema(schema = @Schema(implementation = ProductDTO.class)))) })
    @RequestMapping(value = "/product",
        produces = { "*/*" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<ProductDTO>> allProducts();


    @Operation(summary = "Получить товар по категории", description = "Получить товар по категории", tags={ "product" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "OK", content = @Content(mediaType = "*/*", array = @ArraySchema(schema = @Schema(implementation = ProductDTO.class)))) })
    @RequestMapping(value = "/product/category/{category}",
        produces = { "*/*" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<ProductDTO>> getProductByCategory(@Parameter(in = ParameterIn.PATH, description = "", required=true, schema=@Schema()) @PathVariable("category") String category
);


    @Operation(summary = "Получить товар по ID", description = "Получить товар по ID", tags={ "product" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "OK", content = @Content(mediaType = "*/*", schema = @Schema(implementation = ProductDTO.class))) })
    @RequestMapping(value = "/product/{id}",
        produces = { "*/*" }, 
        method = RequestMethod.GET)
    ResponseEntity<ProductDTO> getProductById(@Min(1L)@Parameter(in = ParameterIn.PATH, description = "", required=true, schema=@Schema(allowableValues={ "1" }, minimum="1"
)) @PathVariable("id") Long id
);


    @Operation(summary = "Получить товар по названию", description = "Получить товар по названию", tags={ "product" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "OK", content = @Content(mediaType = "*/*", array = @ArraySchema(schema = @Schema(implementation = ProductDTO.class)))) })
    @RequestMapping(value = "/product/name/{name}",
        produces = { "*/*" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<ProductDTO>> getProductByName(@Parameter(in = ParameterIn.PATH, description = "", required=true, schema=@Schema()) @PathVariable("name") String name
);


    @Operation(summary = "Удалить товар по ID", description = "Удалить товар по ID", tags={ "product" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "OK") })
    @RequestMapping(value = "/product/{id}",
        method = RequestMethod.DELETE)
    ResponseEntity<Void> removeProductById(@Min(1L)@Parameter(in = ParameterIn.PATH, description = "", required=true, schema=@Schema(allowableValues={ "1" }, minimum="1"
)) @PathVariable("id") Long id
);


    @Operation(summary = "Изменить продукт", description = "Изменить продукт", tags={ "product" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "OK") })
    @RequestMapping(value = "/product/{id}",
        consumes = { "application/json" }, 
        method = RequestMethod.PATCH)
    ResponseEntity<Void> updateProduct(@Min(1L)@Parameter(in = ParameterIn.PATH, description = "", required=true, schema=@Schema(allowableValues={ "1" }, minimum="1"
)) @PathVariable("id") Long id
, @Parameter(in = ParameterIn.DEFAULT, description = "", required=true, schema=@Schema()) @Valid @RequestBody ProductCreationDTO body
);

}